# UsersApi/wwwroot/openapi-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "contact" : {
      "name" : "One Technologies Team"
    },
    "description" : "A simple API to manage users",
    "title" : "User API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local server",
    "url" : "https://localhost:5001"
  } ],
  "tags" : [ {
    "description" : "The user API",
    "name" : "user"
  } ],
  "paths" : {
    "/users" : {
      "get" : {
        "description" : "Returns all users",
        "operationId" : "GetUsers",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "A list of users"
          }
        },
        "summary" : "Get all users",
        "tags" : [ "user" ]
      }
    },
    "/user" : {
      "post" : {
        "description" : "Creates a new user",
        "operationId" : "CreateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "description" : "User to create",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User created"
          }
        },
        "summary" : "Create a new user",
        "tags" : [ "user" ]
      }
    },
    "/user/{id}" : {
      "delete" : {
        "description" : "Deletes a user by id",
        "operationId" : "DeleteUserById",
        "parameters" : [ {
          "description" : "The user id",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "User deleted successfully"
          }
        },
        "summary" : "Delete a user by id",
        "tags" : [ "user" ]
      },
      "get" : {
        "description" : "Returns a user by id",
        "operationId" : "GetUserById",
        "parameters" : [ {
          "description" : "The user id",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "A user"
          }
        },
        "summary" : "Get a user by id",
        "tags" : [ "user" ]
      },
      "put" : {
        "description" : "Updates a user by id",
        "operationId" : "UpdateUserById",
        "parameters" : [ {
          "description" : "The user id",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "description" : "User to update",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User updated"
          }
        },
        "summary" : "Update a user by id",
        "tags" : [ "user" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "example" : {
          "phone" : "phone",
          "name" : "name",
          "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "email" : "email"
        },
        "properties" : {
          "id" : {
            "description" : "The user id",
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "description" : "The user name",
            "type" : "string"
          },
          "email" : {
            "description" : "The user email",
            "type" : "string"
          },
          "phone" : {
            "description" : "The user phone",
            "type" : "string"
          }
        },
        "required" : [ "email", "id", "name", "phone" ],
        "type" : "object"
      },
      "CreateUserRequest" : {
        "example" : {
          "phone" : "phone",
          "name" : "name",
          "email" : "email"
        },
        "properties" : {
          "name" : {
            "description" : "The user name",
            "type" : "string"
          },
          "email" : {
            "description" : "The user email",
            "type" : "string"
          },
          "phone" : {
            "description" : "The user phone",
            "type" : "string"
          }
        },
        "required" : [ "email", "name", "phone" ],
        "type" : "object"
      }
    }
  }
}